Remove leftover unused code
