+++ osu.Game/Screens/OnlinePlay/Matchmaking/Match/BeatmapSelect/MatchmakingSelectPanelBeatmap.cs
@@ -31,6 +31,6 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.Match.BeatmapSelect
 
         private BeatmapCardMatchmakingBeatmapContent? content;
 
-        protected override float AvatarOverlayOffset => base.AvatarOverlayOffset + (content?.AvatarOffset ?? 0);
+        protected override float AvatarOverlayOffset => content?.AvatarOffset ?? 0;
     }
 }
